Factor out BufferStream
