+++ lib/storage/index.js
@@ -22,8 +22,6 @@
 
 var crypto = require('crypto');
 var duplexify = require('duplexify');
-var nodeutil = require('util');
-var stream = require('stream');
 var uuid = require('node-uuid');
 
 /**
@@ -37,6 +35,12 @@ var conn = require('../common/connection.js');
  * @private
  */
 var util = require('../common/util.js');
+
+/**
+ * @type module:common/bufferstream
+ * @private
+ */
+var BufferStream = require('../common/bufferstream.js');
 
 /**
  * Required scopes for Google Cloud Storage API.
@@ -56,31 +60,6 @@ var STORAGE_BASE_URL = 'https://www.googleapis.com/storage/v1/b';
  * @private
  */
 var STORAGE_UPLOAD_BASE_URL = 'https://www.googleapis.com/upload/storage/v1/b';
-
-/**
- * Readable stream implementation to stream the given buffer.
- *
- * @constructor
- *
- * @param {buffer} buffer - The buffer to stream.
- *
- * @private
- */
-function BufferStream(buffer) {
-  stream.Readable.call(this);
-  this.data = buffer;
-}
-
-nodeutil.inherits(BufferStream, stream.Readable);
-
-/**
- * Push the provided buffer to the stream.
- * @private
- */
-BufferStream.prototype._read = function() {
-  this.push(this.data);
-  this.push(null);
-};
 
 /**
  * Google Cloud Storage allows you to store data on Google infrastructure. See
